Renamed TODOListReducer to TODOList_Slice.js.
Wrapped both the sidebar and TODOList components in separate containers in App.js so that both will render nicely. The container holding TODOList component, of className "dashBoardView," will hopefully be the container where we add other components of StudyDash.

For now, we will be storing all of our redux slices within rootReducer.js. This will be our combined redux store.