Fixed a serialization error regarding dueDate attribute of each TODO item.
The reason for this error was that Date objects are not directly serializable. I resolved it by now storing the dueDate attributes of each TODOItem as date strings within the redux store.

I removed the enforced lowercasing of category inputs for TODO items. That was not good, since categories can be capitalized.

I also removed duplication in reducer functions for todoListSlice.js caused by past merges to main branch from others.