Added Calendar view of TODOList
Implemented TODOCalendarView.js to add a Calendar view of the TODO list into the TODOList dashboard view.

Added the necessary useEffect hook so that any changes made to the TODOList in the MongoDB database will be instantly reflected in the calendar view.

Created TODOCalendarIcon.js so that the presence of TODOItems in each date tile of the calendar is represented by rendering an icon for those tiles.

Clicking on the icon will render a popover list of relevant TODOItems for that date tile.

Removed enforced lowercasing of each TODO item’s title attribute in TODOItem.model.js.

NOTE:

There is currently a minor issue with the CSS styling.

Say that I click upon a "Date" tile in the Calendar. At that moment the content within the tile turns solid white. That is fine. Clicking another date tile removes the solid white from the earlier clicked tile.

However, when I click outside of the calendar tiles area, the tile I clicked earlier remains solid white. That is not good. I was expected the tile I clicked earlier to no longer be solid white. I cannot figure out how to resolve this. You guys can decide whether this is a big deal or not.