Fixed disappearing text issue in calendar date tiles
It turns out the issue was that the text colour of each date tile was white. So, when I click away from the Calendar and the background of the previously clicked date tile turns to white, the contents of the tile are no longer legible. I have solved this by setting the tile text colour to black.