Fixed TODOList issues raised during milestone 5 demo
Updated TODOList.css so that the todo items icon is positioned in the top-right corner of the date tile. The date value of a date tile (e.g. the “16” for the “July 16” date tile) is no longer pushed around by the icon.

Created TODOListModule.js to add the calendar view to the TODO sidebar page.

Modified TODOCalendarIcon.js so that the todo list icon colour turns green if all todos for a given date tile of the Calendar are marked as finished.

Modified TODOCalendarView.js so that when we click a date tile once, it correctly shows the orange “selection colour." Clicking the same tile again removes the selection colour.

Modified TODOCalendarView.js so that todo item icons are added to date tiles associated with the due dates of the associated todos.

Modified TODOItem.js so that the time values shown for each todo item card do not include seconds.